# Reduce peak memory usage in forward search