--- a/NEWS.md
+++ b/NEWS.md
@@ -24,6 +24,7 @@ If you read this from a place other than <https://mc-stan.org/projpred/news/inde
 * Output element `p_type` of `project()` has been removed. Instead, output element `const_wdraws_prj` has been added, but its definition is essentially the inverse of former element `p_type` (see the updated documentation of `project()`'s output). This should not be a breaking change for users (as `p_type` was mainly intended for internal use and the new element `const_wdraws_prj` is so, too) but this slightly enhances the cases where `as.matrix.projection()` throws a warning concerning the weights of the projected draws and the cases where `proj_predict()` resamples from the projected draws using argument `nresample_clusters`. (GitHub: #432)
 * Function `do_call()` is deprecated and will be removed in a future release. Where possible, please use direct function calls instead. If this is not possible, please use `do.call()` instead.
 * Improved handling of PSIS-LOO CV warnings. (GitHub: #438)
+* Reduced peak memory usage during forward search. (GitHub: #442)
 
 ## Bug fixes