--- a/R/search.R
+++ b/R/search.R
@@ -32,6 +32,10 @@ search_forward <- function(p_ref, refmodel, nterms_max, verbose = TRUE, opt,
       }
       verb_out(vtxt)
     }
+
+    # Free up some memory:
+    rm(submodls)
+    gc()
   }
 
   # For `solution_terms`, `reduce_models(chosen)` used to be used instead of