--- a/R/search.R
+++ b/R/search.R
@@ -40,6 +40,9 @@ search_forward <- function(p_ref, refmodel, nterms_max, verbose = TRUE, opt,
 
     # Free up some memory:
     rm(submodls)
+    if (getOption("projpred.run_gc", TRUE)) {
+      gc()
+    }
   }
 
   # For `solution_terms`, `reduce_models(chosen)` used to be used instead of